Implement "edit todo" logic
- Added a new "todoControls.js" file to better organize the project.
Future commits will make use of this file to implement the todo removal
functionality.

- Added a new method for getting todos using a todoId. I'm not reusing
the todolist array because that makes the code slightly harder to read.

- Fixed the "editTodo" method returning -1 for the index. The todoId
that is passed to editTodo is of type String so we have to convert it to
a type Number to get the correct value.

- Removed "edit-btn" and "remove-btn" classes from the todo buttons.
Instead, these buttons will make use of a data attribute with no value

- Added a new utility function to format the date for display when
opening the "edit todo" modal.